Ajustes na autenticação
Realizado ajustes na autenticação da entidade Pessoa.

+++ controle/src/main/java/com/estoque/service/PessoaService.java
@@ -1,10 +1,12 @@
 package com.estoque.service;
-
-import java.util.Date;
 import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.estoque.exception.UnauthorizedException;
 import com.estoque.model.Pessoa;
+import com.estoque.model.Enums.TipoPermissao;
 import com.estoque.repository.PessoaRepository;
 
 @Service
@@ -13,21 +15,31 @@ public class PessoaService {
     @Autowired
     private PessoaRepository pessoaRepository;
 
-    public List<Pessoa> buscarTodos() {
+    public List<Pessoa> buscarTodos(TipoPermissao permissao) {
+        if (!permissao.buscarTodos()) {
+            throw new UnauthorizedException("Permissão negada para buscar registros.");
+        }
         return pessoaRepository.findAll();
     }
 
-    public Pessoa inserirPessoa(Pessoa objeto) {
-        objeto.setDataCriacaoPessoa(new Date());
+    public Pessoa inserirPessoa(Pessoa objeto, TipoPermissao permissao) {
+        if (!permissao.inserir()) {
+            throw new UnauthorizedException("Permissão negada para inserir registros.");
+        }
         return pessoaRepository.saveAndFlush(objeto);
     }
 
-    public Pessoa alterarPessoa(Pessoa objeto) {
-        objeto.setDataAtualizacaoPessoa(new Date());
+    public Pessoa alterarPessoa(Pessoa objeto, TipoPermissao permissao) {
+        if (!permissao.alterar()) {
+            throw new UnauthorizedException("Permissão negada para alterar registros.");
+        }
         return pessoaRepository.saveAndFlush(objeto);
     }
 
-    public void excluir(Long id) {
+    public void excluir(Long id, TipoPermissao permissao) {
+        if (!permissao.excluir()) {
+            throw new UnauthorizedException("Permissão negada para excluir registros.");
+        }
         Pessoa objeto = pessoaRepository.findById(id).orElse(null);
         if (objeto != null) {
             pessoaRepository.delete(objeto);
